CH Refactor: Group "Related Top-Performing Posts" files into directory
